// packages/server-nest/src/modules/FinancialStatements/modules/BalanceSheet/BalanceSheet.types.ts
export interface IBalanceSheetQuery extends IFinancialSheetBranchesQuery {
  displayColumnsType: 'total' | 'date_periods';
  displayColumnsBy: string;
  fromDate: Date;
  toDate: Date;
  numberFormat: INumberFormatQuery;
  noneTransactions: boolean;
  noneZero: boolean;
  basis: 'cash' | 'accrual';
  accountIds: number[];

  percentageOfColumn: boolean;
  percentageOfRow: boolean;

  previousPeriod: boolean;
  previousPeriodAmountChange: boolean;
  previousPeriodPercentageChange: boolean;

  previousYear: boolean;
  previousYearAmountChange: boolean;
  previousYearPercentageChange: boolean;
}
// ...
export enum IAccountTransactionsGroupBy {
  Quarter = 'quarter',
  Year = 'year',
  Day = 'day',
  Month = 'month',
  Week = 'week',
}